Refactor web and rest server
Rename files
